refactor: moved all tests to __tests__
